Make organized components consistent with index.js
