+++ src/components/ComponentFactory.js
@@ -4,10 +4,10 @@ import Admonition, { admonitionMap } from './Admonition';
 import Banner from './Banner/Banner';
 import BlockQuote from './BlockQuote';
 import Button from './Button';
-import Card from './Card/Card';
+import Card from './Card';
 import CardGroup from './Card/CardGroup';
 import Chapter from './Chapters/Chapter';
-import Chapters from './Chapters/Chapters';
+import Chapters from './Chapters';
 import Code from './Code/Code';
 import CodeIO from './Code/CodeIO';
 import Cond from './Cond';
@@ -25,10 +25,10 @@ import Extract from './Extract';
 import Field from './FieldList/Field';
 import FieldList from './FieldList/FieldList';
 import Figure from './Figure';
-import Footnote from './Footnote/Footnote';
+import Footnote from './Footnote';
 import FootnoteReference from './Footnote/FootnoteReference';
 import Glossary from './Glossary';
-import GuideNext from './GuideNext/GuideNext';
+import GuideNext from './GuideNext';
 import Heading from './Heading';
 import HorizontalList from './HorizontalList';
 import Image from './Image';
@@ -36,8 +36,8 @@ import Include from './Include';
 import Introduction from './Introduction';
 import Kicker from './Kicker';
 import Line from './LineBlock/Line';
-import LineBlock from './LineBlock/LineBlock';
-import List from './List/List';
+import LineBlock from './LineBlock';
+import List from './List';
 import ListItem from './List/ListItem';
 import ListTable from './ListTable';
 import Literal from './Literal';
@@ -48,7 +48,7 @@ import MongoWebShell from './MongoWebShell';
 import OpenAPI from './OpenAPI';
 import Operation from './Operation';
 import Paragraph from './Paragraph';
-import Procedure from './Procedure/Procedure';
+import Procedure from './Procedure';
 import QuizChoice from './Widgets/QuizWidget/QuizChoice';
 import QuizWidget from './Widgets/QuizWidget/QuizWidget';
 import Reference from './Reference';
@@ -62,7 +62,7 @@ import Strong from './Strong';
 import Subscript from './Subscript';
 import SubstitutionReference from './SubstitutionReference';
 import Superscript from './Superscript';
-import Tabs from './Tabs/Tabs';
+import Tabs from './Tabs';
 import Target from './Target';
 import Text from './Text';
 import Time from './Time';
@@ -71,7 +71,7 @@ import Topic from './Topic';
 import Transition from './Transition';
 import Twitter from './Twitter';
 import VersionModified from './VersionModified';
-import Video from './Video/Video';
+import Video from './Video';
 
 import RoleAbbr from './Roles/Abbr';
 import RoleClass from './Roles/Class';
