+++ src/components/Footnote/index.js
