+++ tests/unit/Card.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { ThemeProvider } from '@emotion/react';
-import Card from '../../src/components/Card/Card';
+import Card from '../../src/components/Card';
 import { theme } from '../../src/theme/docsTheme';
 // data for this component
 import mockData from './data/Card.test.json';
