+++ tests/unit/Footnote.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Footnote from '../../src/components/Footnote/Footnote';
+import Footnote from '../../src/components/Footnote';
 import FootnoteContext from '../../src/components/Footnote/footnote-context';
 
 // data for this component
