+++ tests/unit/List.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import List from '../../src/components/List/List';
+import List from '../../src/components/List';
 
 // data for this component
 import mockData from './data/List.test.json';
