+++ tests/unit/Video.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Video from '../../src/components/Video/Video';
+import Video from '../../src/components/Video';
 
 // data for this component
 import mockData from './data/Video.test.json';
